requirements.txt
# Framework web y extensiones
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
email-validator==1.1.3

# Base de datos y ORM
SQLAlchemy==1.4.23
alembic==1.7.1

# Seguridad y autenticación
python-dotenv==0.19.0
Werkzeug==2.0.1

# Procesamiento de imágenes
Pillow==8.3.1

# Desarrollo y debugging
python-dotenv==0.19.0